refactor: split api tests for specific routes
